feat(memory): land Phase 0-2 as one integration (image route, cascade compaction, memory gate, relevance overhaul)

+++ brain/recovery/source_reader.py
@@ -67,7 +67,7 @@ def read_source_memories(source_dir: Path) -> dict[str, Memory]:
             metadata=_loads_dict(d.get("metadata_json")),
             state=d.get("state") or "active",
             content_snapshot=d.get("content_snapshot"),
-            recall_count=int(d.get("recall_count") or 0),
+            recall_count=float(d.get("recall_count") or 0.0),
             peak_emotion_intensity=float(d.get("peak_emotion_intensity") or 0.0),
         )
     return out
